requirements.txt
KahootPY==0.1.4
colorama==0.4.4
pyfiglet